# apps/api/openapi.routes.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RemnaRay API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/v1/auth/login": {
      "post": {
        "operationId": "adminLogin",
        "responses": {
          "200": {
            "description": "Password challenge"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "423": {
            "description": "Admin locked"
          }
        }
      }
    },
...
    "/api/admin/v1/auth/me": {
      "get": {
        "operationId": "adminMe",
        "responses": {
          "200": {
            "description": "Current admin and CSRF token"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/api/admin/v1/auth/totp": {
      "post": {
        "operationId": "adminTotpLogin",
        "responses": {
          "200": {
            "description": "Authenticated admin session"
          },
          "401": {
            "description": "Invalid TOTP"
          }
        }
      }
    },
...
    "/api/v1/auth/telegram/oidc": {
      "post": {
        "operationId": "telegramOidcLogin",
        "summary": "Telegram OIDC login: verify the id_token and open a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["idToken"],
                "properties": {
                  "idToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ user } and the rr_sid cookie"
          },
          "401": {
            "description": "AUTH_INVALID_SIGNATURE | AUTH_EXPIRED"
          },
          "503": {
            "description": "AUTH_UNAVAILABLE"
          }
        }
      }
    },
    "/api/v1/auth/tg": {
      "post": {
        "operationId": "botLinkSignIn",
        "summary": "The bot's account link, confirmed on its page: body { token }; spends the link and sets rr_sid (R79, L-3)",
        "responses": {
          "204": {
            "description": "Signed in"
          },
          "401": {
            "description": "The link is invalid, expired or spent"
          }
        }
      }
    },
    "/api/v1/auth/tg/preview": {
      "post": {
        "operationId": "botLinkPreview",
        "summary": "Whose account the bot's link opens, without spending it: body { token }",
        "responses": {
          "200": {
            "description": "{ user: { firstName: string | null, username: string | null } }"
          },
          "401": {
            "description": "The link is invalid, expired or spent"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getMe",
        "summary": "UserMe",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      },
      "patch": {
        "operationId": "patchMe",
        "summary": "Update language, email and marketing opt-out",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/r/{code}": {
      "get": {
        "operationId": "followReferralLink",
        "summary": "Stores the referral code for 30 days and redirects to the site",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to /",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}